refactor: create utils, helpers, clean up root src files
