+++ mcps/mcp-with-express/src/constants/config.constants.ts
@@ -1,0 +1,2 @@
+export const NWS_API_BASE = "https://api.weather.gov";
+export const USER_AGENT = "weather-app/1.0";
